refactor(actionType, mapError): util functions for dealing with errors
- `actionType` returns the type of the action: 'request', 'success', 'failure'
- `mapError`: my hypothesis for how to deal with messy server error responses. Use an object to map the server response to what we want to show on the frontend. We can use this to refactor the backend's responses so they are more legible for human consumption
